// src/app/models/vehicles.model.ts
export class VehiclesModel{
	plate: string = "";
	engine: string = "";
	chassis: string = "";
	model: string = "";
	registration_date: string = "";
	seated_passengers: string = "";
	standing_passengers: string = "";
	gross_weight: string = "";
	net_weight: string = "";
	doors: string = "";
	brand: string = "";
	lineup: string = "";
}